Exclude Yusuf_dev's channel)

+++ delta/channel.ts
@@ -4,7 +4,7 @@ const composer = new Composer();
 
 composer.on("message:text", async (ctx: Context): Promise<void> => {
   if (ctx?.message?.from?.username) {
-    if (ctx?.message?.from?.username === "Channel_Bot") {
+    if (ctx?.message?.from?.username === "Channel_Bot" && ctx?.message?.from?.username !== 'Yusuf_dev') {
       await ctx.deleteMessage();
     }
   }
